+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2023-09-06
+### Added
+- Module to create F5 LB for kafka broker(only)
+- Example on how to use this module
+
 ## [1.0.1] - 2021-10-21
 ### Fixed
 - Output variable dns-a-fqdn in module DNS when provision_dns is false